more stuff, need to do is_admin BOOLEAN
