+++ package.json
@@ -12,11 +12,13 @@
     "bcrypt": "^5.1.0",
     "body-parser": "^1.20.2",
     "dotenv": "^16.3.1",
-    "express": "^4.18.2",
-    "jsonwebtoken": "^9.0.1",
-    "pg": "^8.11.2",
+    "express": "^4.19.2",
+    "jsonwebtoken": "^9.0.2",
+    "morgan": "^1.10.0",
+    "pg": "^8.11.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "uuid": "^9.0.1",
     "vite-express": "*",
     "volleyball": "^1.5.1"
   },
@@ -24,7 +26,12 @@
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.10",
     "@vitejs/plugin-react": "^3.0.1",
-    "nodemon": "^2.0.20",
+    "nodemon": "^2.0.22",
     "vite": "^4.0.4"
-  }
+  },
+  "description": "A template for building web applications using the PERN (PostgreSQL, Express.js, React, Node.js) stack.",
+  "main": "vite.config.js",
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
 }
